Define datastructures for `#[cfg]` attribute, move StrippedCfgItem
